+++ NEWS.md
@@ -1,6 +1,9 @@
 
 ## reticulate 1.15 (UNDER DEVELOPMENT)
 
+- `reticulate::use_condaenv()` better handles cases where no
+  matching environment could be found. (#687)
+  
 - `reticulate` gains the `py_ellipsis()` function, used to access
   the Python `Ellipsis` builtin. (#700, @skeydan)